requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
promptpay==0.1.7
qrcode==7.4.2
Pillow==10.4.0
pytest==8.3.3